third_party/WebKit/LayoutTests/fast/events/touch/gesture/touch-gesture-scroll-iframe-not-propagated-expected.txt


This tests that a gesture scroll is not propagated from an iframe to an outer div when the iframe has no remaining scroll offset when the preventPropagation flag is set for the gesture type GestureScrollUpdate.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
first gesture
PASS iframeScrollingElement.scrollHeight - iframeScrollingElement.scrollTop is iframe.clientHeight
PASS movedbox.scrollTop is 0
PASS movedbox.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDocument]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
